feat(ui): change create password to shared component
