+++ src/locales/en/en.json
@@ -661,10 +661,10 @@
       "skip": "Skip"
     },
     "alert": {
-      "text": "You have chosen to skip creating a password. If you change your mind later, you will still be able to create a password.",
+      "text": "Are you sure you want to skip? By doing this, you will be required to use your passcode for any protected operations. You can, of course, set this up later if you prefer.",
       "button": {
-        "confirm": "Continue without a password",
-        "cancel": "Go back and create a password"
+        "confirm": "Yes, skip",
+        "cancel": "Cancel"
       }
     },
     "error": {
@@ -1130,7 +1130,12 @@
     },
     "newpasscode": {
       "title": "New passcode",
-      "description": "Create a new passcode to re-secure your wallet and perform operations"
+      "description": "Create a new passcode to re-secure your wallet and perform operations",
+      "reenterpasscode": "Re-enter passcode"
+    },
+    "newpassword": {
+      "title": "New password",
+      "description": "Create a new operations password so you can access additional wallet features."
     }
   },
   "shareidentifier": {
