+++ src/ui/components/PasswordModule/PasswordModule.scss
@@ -1,0 +1,42 @@
+.password-module {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  min-height: calc(100vh - (5rem + var(--ion-safe-area-top)));
+
+  & > .page-content {
+    & > * {
+      margin: 0 auto 1.56rem;
+    }
+
+    & > h2 {
+      margin: 1rem 0 0.94rem;
+    }
+  }
+
+  @media screen and (min-width: 250px) and (max-width: 370px) {
+    & > .page-content {
+      & > * {
+        margin: 0 auto 1.248rem;
+      }
+
+      & > h2 {
+        margin: 0.5rem 0 0.94rem;
+      }
+    }
+  }
+
+  --background: var(--ion-color-light);
+
+  ion-item.error {
+    margin-bottom: 0;
+  }
+
+  .error-message-placeholder {
+    display: none;
+  }
+
+  .error-message {
+    text-align: left;
+  }
+}
