+++ src/ui/components/PasswordModule/index.ts
@@ -1,0 +1,1 @@
+export * from "./PasswordModule";
