+++ src/ui/pages/CreatePassword/CreatePassword.scss
@@ -1,18 +1,4 @@
 .create-password {
-  --background: var(--ion-color-light);
-
-  ion-item.error {
-    margin-bottom: 0;
-  }
-
-  .error-message-placeholder {
-    display: none;
-  }
-
-  .error-message {
-    text-align: left;
-  }
-
   .page-header {
     ion-toolbar {
       --background: transparent;
