+++ src/ui/pages/CreatePassword/CreatePassword.tsx
@@ -1,81 +1,30 @@
-import { useState } from "react";
+import { useRef } from "react";
 import { i18n } from "../../../i18n";
-import "./CreatePassword.scss";
-import { CustomInput } from "../../components/CustomInput";
-import { ErrorMessage } from "../../components/ErrorMessage";
 import { RoutePath } from "../../../routes";
-import { Agent } from "../../../core/agent/agent";
-import { KeyStoreKeys, SecureStorage } from "../../../core/storage";
+import { getNextRoute } from "../../../routes/nextRoute";
 import { useAppDispatch, useAppSelector } from "../../../store/hooks";
 import {
   getStateCache,
   setCurrentOperation,
 } from "../../../store/reducers/stateCache";
-import { getNextRoute } from "../../../routes/nextRoute";
 import { updateReduxState } from "../../../store/utils";
-import { Alert } from "../../components/Alert";
-import { OperationType } from "../../globals/types";
-import { MiscRecordId } from "../../../core/agent/agent.types";
 import { PageHeader } from "../../components/PageHeader";
+import { PasswordModule } from "../../components/PasswordModule";
 import { ScrollablePageLayout } from "../../components/layout/ScrollablePageLayout";
-import { PageFooter } from "../../components/PageFooter";
-import { passwordStrengthChecker } from "../../utils/passwordStrengthChecker";
-import { PasswordValidation } from "../../components/PasswordValidation";
+import { OperationType } from "../../globals/types";
 import { useAppIonRouter } from "../../hooks";
-import { BasicRecord } from "../../../core/agent/records";
+import "./CreatePassword.scss";
+import { PasswordModuleRef } from "../../components/PasswordModule/PasswordModule.types";
 
 const CreatePassword = () => {
   const pageId = "create-password";
   const stateCache = useAppSelector(getStateCache);
   const ionRouter = useAppIonRouter();
   const dispatch = useAppDispatch();
-  const [createPasswordValue, setCreatePasswordValue] = useState("");
-  const [confirmPasswordValue, setConfirmPasswordValue] = useState("");
-  const [confirmPasswordFocus, setConfirmPasswordFocus] = useState(false);
-  const [createPasswordFocus, setCreatePasswordFocus] = useState(false);
-  const [hintValue, setHintValue] = useState("");
-  const [alertIsOpen, setAlertIsOpen] = useState(false);
-  const createPasswordValueMatching =
-    createPasswordValue.length > 0 &&
-    confirmPasswordValue.length > 0 &&
-    createPasswordValue === confirmPasswordValue;
-  const createPasswordValueNotMatching =
-    createPasswordValue.length > 0 &&
-    confirmPasswordValue.length > 0 &&
-    createPasswordValue !== confirmPasswordValue;
-  const validated =
-    passwordStrengthChecker.validatePassword(createPasswordValue) &&
-    createPasswordValueMatching &&
-    hintValue !== createPasswordValue;
 
-  const handlePasswordInput = (password: string) => {
-    setCreatePasswordValue(password);
-  };
-  const handleClearState = () => {
-    setCreatePasswordValue("");
-    setConfirmPasswordValue("");
-    setHintValue("");
-  };
+  const passwordModuleRef = useRef<PasswordModuleRef>(null);
 
   const handleContinue = async (skipped: boolean) => {
-    // @TODO - foconnor: We should handle errors here and display something to the user as feedback to try again.
-    if (!skipped) {
-      await SecureStorage.set(
-        KeyStoreKeys.APP_OP_PASSWORD,
-        createPasswordValue
-      );
-      if (hintValue) {
-        await Agent.agent.basicStorage.createOrUpdateBasicRecord(
-          new BasicRecord({
-            id: MiscRecordId.OP_PASS_HINT,
-            content: { value: hintValue },
-          })
-        );
-      }
-    } else {
-      await SecureStorage.set(KeyStoreKeys.PASSWORD_SKIPPED, String(true));
-    }
-
     const { nextPath, updateRedux } = getNextRoute(RoutePath.CREATE_PASSWORD, {
       store: { stateCache },
       state: { skipped },
@@ -92,7 +41,6 @@ const CreatePassword = () => {
     );
     dispatch(setCurrentOperation(OperationType.IDLE));
     ionRouter.push(nextPath.pathname, "forward", "push");
-    handleClearState();
   };
 
   return (
@@ -101,7 +49,7 @@ const CreatePassword = () => {
       header={
         <PageHeader
           backButton={true}
-          beforeBack={handleClearState}
+          beforeBack={passwordModuleRef.current?.clearState}
           currentPath={RoutePath.CREATE_PASSWORD}
           progressBar={true}
           progressBarValue={0.5}
@@ -109,83 +57,12 @@ const CreatePassword = () => {
         />
       }
     >
-      <h2 data-testid={`${pageId}-title`}>{i18n.t("createpassword.title")}</h2>
-      <p
-        className="page-paragraph"
-        data-testid={`${pageId}-top-paragraph`}
-      >
-        {i18n.t("createpassword.description")}
-      </p>
-      <CustomInput
-        dataTestId="createPasswordValue"
-        title={`${i18n.t("createpassword.input.first.title")}`}
-        placeholder={`${i18n.t("createpassword.input.first.placeholder")}`}
-        hiddenInput={true}
-        onChangeInput={(password: string) => handlePasswordInput(password)}
-        onChangeFocus={setCreatePasswordFocus}
-        value={createPasswordValue}
-        error={
-          !createPasswordFocus &&
-          !!createPasswordValue.length &&
-          (!passwordStrengthChecker.validatePassword(createPasswordValue) ||
-            !passwordStrengthChecker.isValidCharacters(createPasswordValue))
-        }
-      />
-      {createPasswordValue && (
-        <PasswordValidation password={createPasswordValue} />
-      )}
-      <CustomInput
-        dataTestId="confirm-password-value"
-        title={`${i18n.t("createpassword.input.second.title")}`}
-        placeholder={`${i18n.t("createpassword.input.second.placeholder")}`}
-        hiddenInput={true}
-        onChangeInput={setConfirmPasswordValue}
-        onChangeFocus={setConfirmPasswordFocus}
-        value={confirmPasswordValue}
-        error={
-          !confirmPasswordFocus &&
-          !!confirmPasswordValue.length &&
-          createPasswordValueNotMatching
-        }
-      />
-      {!confirmPasswordFocus &&
-        !!confirmPasswordValue.length &&
-        createPasswordValueNotMatching && (
-        <ErrorMessage
-          message={`${i18n.t("createpassword.error.hasNoMatch")}`}
-        />
-      )}
-      <CustomInput
-        dataTestId="hintValue"
-        title={`${i18n.t("createpassword.input.third.title")}`}
-        placeholder={`${i18n.t("createpassword.input.third.placeholder")}`}
-        onChangeInput={setHintValue}
-        optional={true}
-        value={hintValue}
-        error={!!hintValue.length && hintValue === createPasswordValue}
-      />
-      {!!hintValue.length && hintValue === createPasswordValue && (
-        <ErrorMessage
-          message={`${i18n.t("createpassword.error.hintSameAsPassword")}`}
-        />
-      )}
-
-      <PageFooter
-        pageId={pageId}
-        primaryButtonText={`${i18n.t("createpassword.button.continue")}`}
-        primaryButtonAction={() => handleContinue(false)}
-        primaryButtonDisabled={!validated}
-        tertiaryButtonText={`${i18n.t("createpassword.button.skip")}`}
-        tertiaryButtonAction={() => setAlertIsOpen(true)}
-      />
-      <Alert
-        isOpen={alertIsOpen}
-        setIsOpen={setAlertIsOpen}
-        dataTestId="create-password-alert-skip"
-        headerText={`${i18n.t("createpassword.alert.text")}`}
-        confirmButtonText={`${i18n.t("createpassword.alert.button.confirm")}`}
-        cancelButtonText={`${i18n.t("createpassword.alert.button.cancel")}`}
-        actionConfirm={() => handleContinue(true)}
+      <PasswordModule
+        ref={passwordModuleRef}
+        testId={pageId}
+        title={`${i18n.t("createpassword.title")}`}
+        description={`${i18n.t("createpassword.description")}`}
+        onCreateSuccess={handleContinue}
       />
     </ScrollablePageLayout>
   );
